Since the fonts now use the correct names, the Fontmap file can lose all the extra crap mapping the GS fonts to the normal font names.
Added pdftops filter to makefile install rules.

pdf_main.ps was putting the copyright notice in the document header.  Since
we have both the ESP and Aladdin copyrights in there, this had to be
commented out...

Added check for CONTENT_TYPE environment variable in pstoraster.c; for
PDF files we automagically treat the PS RIP as a PDF to PS converter.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@567 7a7537e8-13f0-0310-91df-b6672ffda945