More config changes, this time to finalize man page generation so that "cat" pages are formatted with numeric, .0, or .z extensions as needed.
All mkdir's are followed by chmod's to ensure that the default umask on
anal OS's does not interfere with a good install.

genlist now creates many HTML files for all available printers; these
are used in place of the 60+ page appendix in the admin guide.

genppd now makes the printer name "Make Model, version" - the comma and
space needed to be swapped...


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@1194 7a7537e8-13f0-0310-91df-b6672ffda945