Update SuSE PAM configuration file.
Reload jobs after reloading configuration files - otherwise the mime type
pointers would be bad.

Add -F option for running the scheduler in the foreground but still
detach from the console and directory.

When running in daemon mode, wait for the child process to start accepting
connections before returning from the parent.

Make sure to set the network interface cache time (so that the cache is
actually used...)


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@2566 7a7537e8-13f0-0310-91df-b6672ffda945