Set the CLASS environment variable when sending jobs to a class.
Preserved class and remote bits for dtype.

Return from the IPP backend immediately (well, after 5 seconds) if the
backend cannot connect to the print server or submit a print job (but
only if we are printing to a class...)

Don't stop the destination queue if the backend failed when printing
to a class.

Update man pages for backend and filter env vars.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@2580 7a7537e8-13f0-0310-91df-b6672ffda945