ppdEmit() and friends now use the RequiresPageRegion attribute (along with the presence of CUPS filters) to determine whether to include PageRegion commands along with InputSlot commands.
Added additional PPD status codes and messages.

Updated ppdOpen() and friends to support standard options outside of
OpenUI/CloseUI, to conform to the PPD spec.

Updated cupstestppd for a better default (concise) report, which shows
pass/fail on the first line and any error and warning conditions after.
The new reports now also reference the corresponding section and pages
in the spec for the failure/warning in question.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@3373 7a7537e8-13f0-0310-91df-b6672ffda945