Add configure options to set the default values of various cupsd.conf settings.
Fix the OSX test package script to update the web interface files before
doing the install into the package directory - that way the revision
number appears.

Generate template/header.tmpl so that the version number is shown correctly
on all pages.

In cupsdStartBrowsing(), decouple the socket creation from the socket
"broadcast" and "close on exec" options, since launchd will be providing
the BrowseSocket for us on MacOS X.



git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@5008 7a7537e8-13f0-0310-91df-b6672ffda945