The scheduler now delays writing config and state files to reduce disk activity (STR #2684)
Added cupsdCleanDirty, cupsdMarkDirty, and cupsdSetBusy functions in the
scheduler for managing the writing of all config and state files to disk.

Instead of calling cupsdSave* or cupsdWritePrintcap directly, now we call
cupsdMarkDirty and (for jobs) set the "dirty" flag on the job.  The main
loop then calls cupsdCleanDirty periodically as needed.

In addition, we now track the jobs that are currently printing in the
PrintingJobs array - this will allow for better power management support
in the future through the cupsdSetBusy function.



git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@7468 7a7537e8-13f0-0310-91df-b6672ffda945