Scheduler crash due to unknown circumstances leading to NULL policy pointer (STR #4378)
This should never happen, but could be due to a patch applied for Debian/Ubuntu.

This change logs the occurrence and uses the default policy if the policy is
NULL.  If DefaultPolicyPtr is NULL (something that definitely should never
happen when cupsd is accepting client connections) then we will still crash.



git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11681 a1ca3aef-8c08-0410-bb20-df032aa958be