Add --list-filters options to cupsfilter (STR #4325)
Also update cupsfilter to look for cups-files.conf instead of cupsd.conf (this
was never updated after the security changes to use cups-files.conf)

Update the man page to use the "modern" markup and sections.

Prep for future cups(1) man page that provides an overview of CUPS and the
standard environment.

Update man page converter to not add whitespace before punctuation.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11770 a1ca3aef-8c08-0410-bb20-df032aa958be