The ipptool program now supports collection attributes with multiple values (<rdar://problem/15355124>)
Add a test file for FaxOut.

Update the get_token function to allow {, }, and "," to be values of their own.

Update the collection code to allow multiple top-level collection values (still
no support for embedding member attributes with multiple collection values,
however we don't need to test that right now...)


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11802 a1ca3aef-8c08-0410-bb20-df032aa958be