More man page modernization.
Combine and simplify documentation of cupsd helper programs into a single man
page.

Document the client.conf file as deprecated.

Document the classes.conf, printers.conf, and subscriptions.conf files as
subject to be changed and not to be hand-edited.

Remove separate HTML documentation for client.conf - the man page is now "good
enough"...

Fix usage comment at the top of cups-exec.c.



git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11817 a1ca3aef-8c08-0410-bb20-df032aa958be