configure: allow explicit disabling of Java/PHP/Perl/Python support
Currently, when configure is called with --with-java or --with-java=,
auto-detection is performed. When called with --without-java, HAVE_JAVA
is still being defined. This is unfortunate when cross-compiling for
embedded systems, we would end up with Java on the host being
auto-detected (which is not what we have on the target), or with a wrong
HAVE_JAVA define. We need a way to explicitely disable scripting support
for all supported languages.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>