Let the PPD Generator read color spaces and color depth correctly from IPP data
The PPD generator did not support all combinations of color spaces and
color depths when reading the information from the answer to the
get-printer-attributes IPP request. especially if in Apple Raster a
range of color depths is given for a color space, like "W8-16",
"ADOBERGB24-48", ... Now support for a color space in any color depth
is recognized and the highest depth available chosen for each color
space.