Extract reading dests from lpoptions to separate function
Move the code to read dests from global and user lpoptions
files to a separate function 'cups_get_dests_from_lpoptions'.
This is to avoid duplication, in particular since the same
functionality will be used again in a follow-up patch.

Move the related comment in 'cups_enum_dests' to the place
where the lpoptions files are actually evaluated.