Changed make_device_uri() that it will look for a serial number in the USB device data first before it tries to find the data in the IEEE1284 string.
Existing code also was possibly not likely to work
becausae the devdesc struct was not initialised with
libusb_get_device_descriptor() before _get'ing the
iSerialNumber.

Code now permits the use of multiple printers of the
same model/type even with the same IEEE1284 data SERN
which was the case with the DYMO LabelWriter 450 units.