Fix for #5805 -  number of copies
With that fix, the ppd file that is created by installing a ippeverywhere printer contains correctly "True" and the number of copies is correctly printed. Without that fix, that boot value in the ppd is "true" and as a result the number of copies is ignored and it always prints a single copy.