fix: fix infinite loop waiting for usb printer to be available
Waiting forever for the USB device to be available is not reliable
solution. Sometimes it is not possible to impose a policy of maximum
waiting time for the job to be completed in the scheduler, without
knowing if the process is actually being carried out correctly or not.
Because of this, a maximum number of tries is introduced to make sure
the backend process ends in a failure condition if the USB device is
connected.