Change cupsd.conf to force authentication when accessing administration
With the current default configuration, someone can request to Find New
Printers without being authenticated and with no way to be authenticated.
Without authentication cups does not detects printers on the network.
The person might not realize that it's because authentication is missing.
It seems natural to always offer to authenticate when accessing the
administration page. See https://github.com/apple/cups/issues/6076 .